Fix for Python3 and Rest-Operations. Updated package.xml
